// generated/api-client/models/create-event-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NestJS Authentication API
 * A comprehensive NestJS authentication API with PostgreSQL & Prisma.  ### Features - JWT Authentication with Access & Refresh tokens - Google OAuth2 integration - Local authentication (email/password) - Cookie-based token storage - Protected routes with guards  ### Authentication This API uses JWT tokens stored in HTTP-only cookies for authentication.  1. Local Authentication 2. Google OAuth
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface CreateEventDto
 */
export interface CreateEventDto {
    /**
     * Event title
     * @type {string}
     * @memberof CreateEventDto
     */
    'title': string;
    /**
     * Event type
     * @type {string}
     * @memberof CreateEventDto
     */
    'type'?: CreateEventDtoTypeEnum;
    /**
     * Event image URL
     * @type {string}
     * @memberof CreateEventDto
     */
    'imageUrl'?: string;
    /**
     * Brief event description
     * @type {string}
     * @memberof CreateEventDto
     */
    'description'?: string;
    /**
     * Detailed event description and additional information
     * @type {string}
     * @memberof CreateEventDto
     */
    'details'?: string;
    /**
     * Event start date and time (ISO 8601 format)
     * @type {string}
     * @memberof CreateEventDto
     */
    'startDate': string;
    /**
     * Event end date and time (ISO 8601 format)
     * @type {string}
     * @memberof CreateEventDto
     */
    'endDate'?: string;
    /**
     * Event tags for categorization
     * @type {Array<Array>}
     * @memberof CreateEventDto
     */
    'tags'?: Array<Array>;
    /**
     * Minimum attendance duration in minutes required for certification
     * @type {number}
     * @memberof CreateEventDto
     */
    'minimumAttendanceMinutes'?: number;
}

export const CreateEventDtoTypeEnum = {
    IN_PERSON: 'IN_PERSON',
    VIRTUAL: 'VIRTUAL',
    HYBRID: 'HYBRID',
    OTHERS: 'OTHERS'
} as const;

export type CreateEventDtoTypeEnum = typeof CreateEventDtoTypeEnum[keyof typeof CreateEventDtoTypeEnum];