// generated/api-client/models/update-event-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NestJS Authentication API
 * A comprehensive NestJS authentication API with PostgreSQL & Prisma.  ### Features - JWT Authentication with Access & Refresh tokens - Google OAuth2 integration - Local authentication (email/password) - Cookie-based token storage - Protected routes with guards  ### Authentication This API uses JWT tokens stored in HTTP-only cookies for authentication.  1. Local Authentication 2. Google OAuth
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface UpdateEventDto
 */
export interface UpdateEventDto {
    /**
     * Event title
     * @type {string}
     * @memberof UpdateEventDto
     */
    'title'?: string;
    /**
     * Event type
     * @type {string}
     * @memberof UpdateEventDto
     */
    'type'?: UpdateEventDtoTypeEnum;
    /**
     * Event image URL
     * @type {string}
     * @memberof UpdateEventDto
     */
    'imageUrl'?: string;
    /**
     * Brief event description
     * @type {string}
     * @memberof UpdateEventDto
     */
    'description'?: string;
    /**
     * Detailed event description and additional information
     * @type {string}
     * @memberof UpdateEventDto
     */
    'details'?: string;
    /**
     * Event start date and time (ISO 8601 format)
     * @type {string}
     * @memberof UpdateEventDto
     */
    'startDate'?: string;
    /**
     * Event end date and time (ISO 8601 format)
     * @type {string}
     * @memberof UpdateEventDto
     */
    'endDate'?: string;
    /**
     * Event tags for categorization
     * @type {Array<Array>}
     * @memberof UpdateEventDto
     */
    'tags'?: Array<Array>;
    /**
     * Minimum attendance duration in minutes required for certification
     * @type {number}
     * @memberof UpdateEventDto
     */
    'minimumAttendanceMinutes'?: number;
}

export const UpdateEventDtoTypeEnum = {
    IN_PERSON: 'IN_PERSON',
    VIRTUAL: 'VIRTUAL',
    HYBRID: 'HYBRID',
    OTHERS: 'OTHERS'
} as const;

export type UpdateEventDtoTypeEnum = typeof UpdateEventDtoTypeEnum[keyof typeof UpdateEventDtoTypeEnum];